Move rotate and byteswap intrinsics out of /xcpu to /intrinsics root dir.
