+++ builds/msvc/vs2022/libbitcoin-system/libbitcoin-system.vcxproj
@@ -353,15 +353,15 @@
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\arm\arm.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\arm\functional.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\arm\sha.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\byteswap.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\haves.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\intrinsics.hpp" />
-    <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\byteswap.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\rotate.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\cpuid.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\defines.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\functional_128.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\functional_256.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\functional_512.hpp" />
-    <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\rotate.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\sha.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\intrinsics\xcpu\xcpu.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\system\literals.hpp" />
