+++ include/bitcoin/system.hpp
@@ -129,18 +129,18 @@
 #include <bitcoin/system/hash/sha/sha160.hpp>
 #include <bitcoin/system/hash/sha/sha256.hpp>
 #include <bitcoin/system/hash/sha/sha512.hpp>
+#include <bitcoin/system/intrinsics/byteswap.hpp>
 #include <bitcoin/system/intrinsics/haves.hpp>
 #include <bitcoin/system/intrinsics/intrinsics.hpp>
+#include <bitcoin/system/intrinsics/rotate.hpp>
 #include <bitcoin/system/intrinsics/arm/arm.hpp>
 #include <bitcoin/system/intrinsics/arm/functional.hpp>
 #include <bitcoin/system/intrinsics/arm/sha.hpp>
-#include <bitcoin/system/intrinsics/xcpu/byteswap.hpp>
 #include <bitcoin/system/intrinsics/xcpu/cpuid.hpp>
 #include <bitcoin/system/intrinsics/xcpu/defines.hpp>
 #include <bitcoin/system/intrinsics/xcpu/functional_128.hpp>
 #include <bitcoin/system/intrinsics/xcpu/functional_256.hpp>
 #include <bitcoin/system/intrinsics/xcpu/functional_512.hpp>
-#include <bitcoin/system/intrinsics/xcpu/rotate.hpp>
 #include <bitcoin/system/intrinsics/xcpu/sha.hpp>
 #include <bitcoin/system/intrinsics/xcpu/xcpu.hpp>
 #include <bitcoin/system/machine/interpreter.hpp>
