+++ include/bitcoin/system/intrinsics/byteswap.hpp
@@ -16,8 +16,8 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef LIBBITCOIN_SYSTEM_INTRINSICS_XCPU_BYTESWAP_HPP
-#define LIBBITCOIN_SYSTEM_INTRINSICS_XCPU_BYTESWAP_HPP
+#ifndef LIBBITCOIN_SYSTEM_INTRINSICS_BYTESWAP_HPP
+#define LIBBITCOIN_SYSTEM_INTRINSICS_BYTESWAP_HPP
 
 #include <bitcoin/system/define.hpp>
