+++ include/bitcoin/system/intrinsics/intrinsics.hpp
@@ -20,7 +20,9 @@
 #define LIBBITCOIN_SYSTEM_INTRINSICS_INTRINSICS_HPP
 
 #include <bitcoin/system/intrinsics/arm/arm.hpp>
-#include <bitcoin/system/intrinsics/xcpu/xcpu.hpp>
+#include <bitcoin/system/intrinsics/byteswap.hpp>
 #include <bitcoin/system/intrinsics/haves.hpp>
+#include <bitcoin/system/intrinsics/rotate.hpp>
+#include <bitcoin/system/intrinsics/xcpu/xcpu.hpp>
 
 #endif
